+++ .prettierrc
@@ -1,1 +1,2 @@
-{ "semi": false, "singleQuote": true, "tabWidth": 2 }
+{ "semi": false, "singleQuote": true, "tabWidth": 2, "printWidth": 100, "bracketSpacing": true
+}
