+++ public/index.html
@@ -30,5 +30,6 @@
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
     <script src="/public/main.min.js"></script>
+
   </body>
 </html>
