+++ src/App.jsx
@@ -1,20 +1,20 @@
 import React from 'react'
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
 import { makeStyles } from '@material-ui/core/styles'
-import mainTheme from './theme'
-import NavBar from './components/NavBar'
-import Portfolio from './pages/Portfolio'
-import AboutUs from './pages/AboutUs'
-import Services from './pages/Services'
-import Contact from './pages/Contact'
-import BookCall from './pages/BookCall'
-import Home from './pages/Home'
-import ContentWriting from './pages/ContentWriting'
-import Seo from './pages/Seo'
-import MobileApp from './pages/MobileApp'
-import Ecommerce from './pages/Ecommerce'
-import WebDesign from './pages/WebDesign'
-import WebMaintanice from './pages/WebMaintanice'
+// import mainTheme from './theme'
+import NavBar from '../src/components/layout/NavBar'
+import Portfolio from '../src/components/pages/Portfolio'
+import AboutUs from '../src/components/pages/AboutUs'
+import Services from '../src/components/pages/Services'
+import Contact from '../src/components/pages/Contact'
+import BookCall from '../src/components/pages/BookCall'
+import Home from '../src/components/pages/Home'
+import ContentWriting from '../src/components/pages/ContentWriting'
+import Seo from '../src/components/pages/Seo'
+import MobileApp from '../src/components/pages/MobileApp'
+import Ecommerce from '../src/components/pages/Ecommerce'
+import WebDesign from '../src/components/pages/WebDesign'
+import WebMaintanice from '../src/components/pages/WebMaintanice'
 
 const useStyles = makeStyles(() => ({
   root: {
