+++ src/components/index.js
@@ -1,0 +1,54 @@
+//============================================
+// Layouts
+//============================================
+
+export { default as NavBar } from '../components/layout/NavBar';
+export { default as Footer } from '../components/layout/Footer';
+
+//============================================
+// Sections
+//============================================
+
+export { default as MainDisplay } from '../components/sections/MainDisplay';
+export { default as ClientTest } from '../components/sections/ClientTest';
+export { default as ContentSection } from '../components/sections/ContentSection';
+export { default as FaqServices } from '../components/sections/FaqServices';
+export { default as GetInTouch } from '../components/sections/GetInTouch';
+export { default as GetStarted } from '../components/sections/GetStarted';
+export { default as HmServices } from '../components/sections/HmServices';
+export {default as MissionStatement} from '../components/sections/MissionStatement';
+export { default as OneQuestion } from '../components/sections/OneQuestion';
+export { default as Questions } from '../components/sections/Questions';
+export { default as Resources } from '../components/sections/Resources';
+export { default as SectionHeader } from '../components/sections/SectionHeader';
+export { default as ServicesMain } from '../components/sections/ServicesMain';
+export { default as ServicesOffered } from '../components/sections/ServicesOffered';
+export { default as SimpleSlide } from '../components/sections/SimpleSlide';
+export { default as TechUsed } from '../components/sections/TechUsed';
+export { default as WhatWeDo } from '../components/sections/WhatWeDo';
+export { default as WhoWeAre } from '../components/sections/WhoWeAre';
+
+//============================================
+// Pages
+//============================================
+
+export { default as AboutUs } from '../components/pages/AboutUs';
+export { default as BookCall } from '../components/pages/BookCall';
+export { default as Contact } from '../components/pages/Contact';
+export { default as ContentWriting } from '../components/pages/ContentWriting';
+export { default as Ecommerce } from '../components/pages/Ecommerce';
+export { default as Home } from '../components/pages/Home';
+export { default as MobileApp } from '../components/pages/MobileApp';
+export { default as Portfolio } from '../components/pages/Portfolio';
+export { default as Seo } from '../components/pages/Seo';
+export { default as Services } from '../components/pages/Services';
+export { default as WebDesign } from '../components/pages/WebDesign';
+export { default as WebMaintanice } from '../components/pages/WebMaintanice';
+
+//============================================
+// Forms
+//============================================
+
+export { default as ContactForm } from '../components/forms/ContactForm';
+
+
