+++ src/components/pages/ContentWriting.jsx
@@ -1,6 +1,5 @@
 import React from 'react'
-import ContentSection from '../components/ContentSection'
-import Footer from '../components/Footer'
+import { ContentSection, Footer } from '../index'
 
 function ContentWriting() {
   return (
