+++ src/components/pages/MobileApp.jsx
