+++ src/components/sections/ClientTest.jsx
@@ -1,12 +1,9 @@
 import React from 'react'
 import { makeStyles, useTheme } from '@material-ui/core/styles'
-import mainTheme from '../theme'
-import { Container, Hidden, Grid } from '@material-ui/core'
-import Paper from '@material-ui/core/Paper'
-import Typography from '@material-ui/core/Typography'
+import mainTheme from '../../theme'
+import { Container, Hidden, Grid, Typography, Paper, Avatar  } from '@material-ui/core'
 import SwipeableViews from 'react-swipeable-views'
 import { autoPlay } from 'react-swipeable-views-utils'
-import Avatar from '@material-ui/core/Avatar'
 import FormatQuoteRoundedIcon from '@material-ui/icons/FormatQuoteRounded'
 import StarRoundedIcon from '@material-ui/icons/StarRounded'
 import SectionHeader from './SectionHeader'
