+++ src/components/sections/GetStarted.jsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import { makeStyles } from '@material-ui/core/styles'
-import mainTheme from '../theme'
+import mainTheme from '../../theme'
 import Paper from '@material-ui/core/Paper'
 import { Container, Typography, Grid, Hidden } from '@material-ui/core'
-import SectionHeader from '../components/SectionHeader'
+import SectionHeader from './SectionHeader'
 import BeenhereIcon from '@material-ui/icons/Beenhere'
 import Button from '@material-ui/core/Button'
 import ArrowRightAltIcon from '@material-ui/icons/ArrowRightAlt'
