+++ src/components/sections/Resources.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { makeStyles } from '@material-ui/core/styles'
 import { Grid, Container } from '@material-ui/core'
-import mainTheme from '../theme'
-import SectionHeader from '../components/SectionHeader'
+import mainTheme from '../../theme'
+import SectionHeader from './SectionHeader'
 import Card from '@material-ui/core/Card'
 import CardActionArea from '@material-ui/core/CardActionArea'
 import CardActions from '@material-ui/core/CardActions'
