+++ tests/NavBar.test.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import testComponent from './testComponent'
 
-import NavBar from '../src/components/NavBar'
+import NavBar from '../src/components/layout/NavBar'
 
 testComponent(<NavBar />, 'Navbar')
